Working on the Zapper, but I can't seem to figure out why acos() returns a ridiculously large number, it shouldn't overflow. Confused.

+++ src/Zapper.hpp
@@ -6,7 +6,7 @@
 class Zapper {
 protected:
 	GLfloat location[4][2];	// Original location of the gun
-	GLint rotation; // Current rotation to point at mouse
+	GLfloat rotation; // Current rotation to point at mouse
 public:
 	Zapper(); // Zapper constructor
 	void draw(); // Drawing the Zapper
